WIP: Spot Synchronization Fault
OSS-1-RENAME

**Excessive ptable lock**

Around line 560 of this proc.c, we still hold the ptable lock when switching
a process.

Implementation of xv6 disables all kinds of CPU, including T_SYSCALL, of
course.

During the normal initialization sequence of xv6, 'initcode' process raises a fork
system call to create 'init' process, but the system call interrupt is
ignored, and the initcode process then indefinitely try make the same
attemp.

To-Do
1. Release the lock before switching context.
2. When a process is terminated, it might be possible for a new process
   to hold the same pid with the terminated one. This main scheduler
only takes account of pid to check if the process is serviced
earilier,leaving a potential hazard. Consider checking more properties.
Possibly the creation time of the process.